Add styles SignUpLogin Components
