--- src/components/Signup/signup_ui.js
@@ -1,162 +1,0 @@
-import React, { useState } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { useHistory } from "react-router-dom";
-import { signUp } from "../../actions/users";
-
-// Material UI imports
-import Avatar from "@material-ui/core/Avatar";
-import Button from "@material-ui/core/Button";
-import CssBaseline from "@material-ui/core/CssBaseline";
-import TextField from "@material-ui/core/TextField";
-import Link from "@material-ui/core/Link";
-import Grid from "@material-ui/core/Grid";
-// import Box from "@material-ui/core/Box";
-import Typography from "@material-ui/core/Typography";
-import { makeStyles } from "@material-ui/core/styles";
-import Container from "@material-ui/core/Container";
-
-const useStyles = makeStyles(theme => ({
-  paper: {
-    marginTop: theme.spacing(8),
-    display: "flex",
-    flexDirection: "column",
-    alignItems: "center"
-  },
-  avatar: {
-    margin: theme.spacing(1),
-    backgroundColor: theme.palette.secondary.main
-  },
-  form: {
-    width: "100%",
-    marginTop: theme.spacing(3)
-  },
-  submit: {
-    margin: theme.spacing(3, 0, 2)
-  }
-}));
-
-export default function SignUp() {
-  const classes = useStyles();
-  const dispatch = useDispatch();
-  const history = useHistory();
-  const [state, setState] = useState({
-    username: "",
-    email: "",
-    password: ""
-  });
-
-  // const user = useSelector(state => state.user).userCreated;
-  // console.log("user", user)
-
-  // const handleChange = event => {
-  //   const {value} = event.target.value;
-  //   setState({
-  //     ...state,
-  //     [event.target.name]: value
-  //   });
-  // };
-
-  const handleChange = event => {
-    const { name, value } = event.target;
-    setState(previousValue => ({
-      ...previousValue,
-      [name]: value
-    }));
-  };
-
-  const handleSubmit = event => {
-    event.preventDefault();
-    dispatch(signUp(state.username, state.email, state.password));
-    setState({ username: "", email: "", password: "" });
-    // history.push("/login");
-  };
-
-  const isUserCreated = useSelector(state => state.user.userCreated);
-
-  if (isUserCreated) {
-    setTimeout(() => {
-      history.push("/login");
-    }, 500);
-    return <p>Account created!</p>;
-  }
-
-  return (
-    <>
-      <Container component="main" maxWidth="xs">
-        <CssBaseline />
-        <div className={classes.paper}>
-          <Avatar className={classes.avatar}>
-            {/* <LockOutlinedIcon /> */}
-          </Avatar>
-          <Typography component="h1" variant="h5">
-            Sign up
-          </Typography>
-          <form className={classes.form} noValidate onSubmit={handleSubmit}>
-            <Grid container spacing={2}>
-              <Grid item xs={12}>
-                <TextField
-                  value={state.username}
-                  onChange={handleChange}
-                  // autoComplete="fname"
-                  name="username"
-                  variant="outlined"
-                  required
-                  fullWidth
-                  id="firstName"
-                  label="Username"
-                  type="text"
-                  autoFocus
-                />
-              </Grid>
-              <Grid item xs={12}>
-                <TextField
-                  value={state.email}
-                  onChange={handleChange}
-                  variant="outlined"
-                  required
-                  fullWidth
-                  id="email"
-                  label="Email Address"
-                  name="email"
-                  type="email"
-                />
-              </Grid>
-              <Grid item xs={12}>
-                <TextField
-                  value={state.password}
-                  onChange={handleChange}
-                  variant="outlined"
-                  required
-                  fullWidth
-                  name="password"
-                  label="Password"
-                  type="password"
-                  id="password"
-                />
-              </Grid>
-            </Grid>
-            <Button
-              type="submit"
-              fullWidth
-              variant="contained"
-              color="primary"
-              className={classes.submit}
-              // onClick={() =>
-              //   dispatch(signUp(state.username, state.email, state.password))
-              // }
-            >
-              Sign Up
-            </Button>
-            <Grid container justify="flex-end">
-              <Grid item>
-                <Link href="#" variant="body2">
-                  Already have an account? Sign in
-                </Link>
-              </Grid>
-            </Grid>
-          </form>
-        </div>
-      </Container>
-    </>
-  );
-}
